Initial work on running MTCNN
I ran into the problem that opencv has troubles AGAIN.  Will either try
tomorrow or ask Jip to continue this.  The code doesn't load the images
yet anyway either. Which means that it's obviously not in a finished
state. I guess the setup for the argument parser can be used though!

Signed-off-by: 12690368 <zirk.seljee@student.uva.nl>